requirements.txt
python-dotenv
openai
tiktoken
jinja2
pandas
tavily-python